+++ src/Digdir.Domain.Dialogporten.GraphQL/ServiceCollectionExtensions.cs
@@ -1,5 +1,6 @@
 using Digdir.Domain.Dialogporten.GraphQL.EndUser;
 using Digdir.Domain.Dialogporten.GraphQL.EndUser.DialogById;
+using Digdir.Domain.Dialogporten.GraphQL.EndUser.MutationTypes;
 using Digdir.Domain.Dialogporten.GraphQL.EndUser.SearchDialogs;
 using Digdir.Domain.Dialogporten.Infrastructure.Persistence;
 
@@ -17,11 +18,13 @@ public static class ServiceCollectionExtensions
             .RegisterDbContext<DialogDbContext>()
             .AddDiagnosticEventListener<ApplicationInsightEventListener>()
             .AddQueryType<Queries>()
+            .AddMutationType<Mutations>()
             .AddType<DialogByIdDeleted>()
             .AddType<DialogByIdNotFound>()
             .AddType<DialogByIdForbidden>()
             .AddType<SearchDialogValidationError>()
             .AddType<SearchDialogForbidden>()
+            .AddType<SetSystemLabelEntityNotFound>()
             .AddInstrumentation()
             .InitializeOnStartup()
             .Services;