+++ CHANGELOG.md
@@ -1,5 +1,39 @@
 # Changelog
 
+## [1.21.0](https://github.com/digdir/dialogporten/compare/v1.20.2...v1.21.0) (2024-10-03)
+
+
+### Features
+
+* basic label implementation to hide dialogs ([#1192](https://github.com/digdir/dialogporten/issues/1192)) ([ee90c68](https://github.com/digdir/dialogporten/commit/ee90c6806bf0b394d9062612f5554a4d02616ab4))
+
+
+### Bug Fixes
+
+* **webAPI:** Broken mapping when creating Transmissions in an Update ([#1221](https://github.com/digdir/dialogporten/issues/1221)) ([6e7dfe4](https://github.com/digdir/dialogporten/commit/6e7dfe461eb1841bbdd1dd721fab87e7b609756c))
+
+## [1.20.2](https://github.com/digdir/dialogporten/compare/v1.20.1...v1.20.2) (2024-10-02)
+
+
+### Bug Fixes
+
+* (webAPI): Add revision to search dto (ServiceOwner) ([#1216](https://github.com/digdir/dialogporten/issues/1216)) ([3b6d130](https://github.com/digdir/dialogporten/commit/3b6d130bb117fa8d3e0a183474c9bd60e377abb7))
+* **graphQL:** GraphQL subscription not notified on DialogActivityCreated ([#1187](https://github.com/digdir/dialogporten/issues/1187)) ([f28e291](https://github.com/digdir/dialogporten/commit/f28e291bdba7cf3cc94cf0de84fcc12e781d3abb))
+
+## [1.20.1](https://github.com/digdir/dialogporten/compare/v1.20.0...v1.20.1) (2024-10-02)
+
+
+### Bug Fixes
+
+* Add separate settings for parties cache, don't cache invalid response from Altinn 2 ([#1194](https://github.com/digdir/dialogporten/issues/1194)) ([dbb79dc](https://github.com/digdir/dialogporten/commit/dbb79dc26cefc5f28c21a738f39199c36a49438f))
+
+## [1.20.0](https://github.com/digdir/dialogporten/compare/v1.19.0...v1.20.0) (2024-09-30)
+
+
+### Features
+
+* **GraphQL:** Add DialogToken requirement for subscriptions ([#1124](https://github.com/digdir/dialogporten/issues/1124)) ([651ca62](https://github.com/digdir/dialogporten/commit/651ca62fdec02dec48b674b80acf52737036cf13))
+
 ## [1.19.0](https://github.com/digdir/dialogporten/compare/v1.18.1...v1.19.0) (2024-09-24)